final commit

+++ src/main/java/com/skypro/calculator/controller/CalculatorController.java
@@ -21,23 +21,23 @@ public class CalculatorController {
     }
 
     @GetMapping("/plus")
-    public String plus(@RequestParam("num1") int num1, @RequestParam("num2") int num2) {
-        int result = calculatorService.plus(num1, num2);
+    public String plus(@RequestParam("num1") double num1, @RequestParam("num2") double num2) {
+        double result = calculatorService.plus(num1, num2);
         return num1 + "+" + num2 + "=" + result;
     }
     @GetMapping("/minus")
-    public String minus(@RequestParam("num1") int num1, @RequestParam("num2") int num2) {
-        int result = calculatorService.minus(num1, num2);
+    public String minus(@RequestParam("num1") double num1, @RequestParam("num2") double num2) {
+        double result = calculatorService.minus(num1, num2);
         return num1 + "-" + num2 + "=" + result;
     }
     @GetMapping("/multiply")
-    public String multiply(@RequestParam("num1") int num1, @RequestParam("num2") int num2) {
-        int result = calculatorService.multiply(num1, num2);
+    public String multiply(@RequestParam("num1") double num1, @RequestParam("num2") double num2) {
+        double result = calculatorService.multiply(num1, num2);
         return num1 + "*" + num2 + "=" + result;
     }
     @GetMapping("/divide")
-    public String divide(@RequestParam("num1") int num1, @RequestParam("num2") int num2) {
-        int result = calculatorService.divide(num1, num2);
+    public String divide(@RequestParam("num1") double num1, @RequestParam("num2") double num2) {
+        double result = calculatorService.divide(num1, num2);
         return num1 + "/" + num2 + "=" + result;
     }
 }
