+++ src/main/java/com/skypro/calculator/service/CalculatorService.java
@@ -4,19 +4,19 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CalculatorService {
-    public int plus(int num1, int num2) {
+    public double plus(double num1, double num2) {
         return num1 + num2;
     }
 
-    public int multiply(int num1, int num2) {
+    public double multiply(double num1, double num2) {
         return num1 * num2;
     }
 
-    public int minus(int num1, int num2) {
+    public double minus(double num1, double num2) {
         return num1 - num2;
     }
 
-    public int divide(int num1, int num2) {
+    public double divide(double num1, double num2) {
         return num1 / num2;
     }
 }
